+++ packages/widget/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.5](https://github.com/synapsecns/sanguine/compare/@synapsecns/widget@0.6.4...@synapsecns/widget@0.6.5) (2024-08-08)
+
+**Note:** Version bump only for package @synapsecns/widget
+
+
+
+
+
 ## [0.6.4](https://github.com/synapsecns/sanguine/compare/@synapsecns/widget@0.6.3...@synapsecns/widget@0.6.4) (2024-08-06)
 
 **Note:** Version bump only for package @synapsecns/widget